starting LD51 game
